pip-requirements.txt
xmltodict==0.12.0
suds-py3==1.4.0.0
ngram==3.3.2
StringDist==1.0.9
beautifulsoup4==4.9.0
plyvel==1.2.0
more-itertools==8.2.0
retry==0.9.2
pydecor==2.0.0
